Register menu items for StereoToMono
