+++ src/effects/StereoToMono.cpp
@@ -227,3 +227,52 @@ bool EffectStereoToMono::IsHiddenFromMenus() const
 {
    return true;
 }
+
+// Attach a menu item
+#include "CommonCommandFlags.h"
+#include "PluginManager.h"
+#include "commands/CommandManager.h"
+#include "effects/EffectManager.h"
+#include "effects/EffectUI.h"
+
+namespace {
+void OnStereoToMono(const CommandContext &context)
+{
+   EffectUI::DoEffect(
+      EffectManager::Get().GetEffectByIdentifier(wxT("StereoToMono")),
+      context,
+      EffectManager::kConfigured);
+}
+
+using namespace MenuTable;
+BaseItemSharedPtr MenuItem()
+{
+   using Options = CommandManager::Options;
+
+   static BaseItemSharedPtr menu {
+   // Delayed evaluation
+   // Stereo to Mono is an oddball command that is also subject to control
+   // by the plug-in manager, as if an effect.  Decide whether to show or
+   // hide it.
+   Items( "",
+      [](AudacityProject&) -> BaseItemPtr {
+         const PluginID ID =
+            EffectManager::Get().GetEffectByIdentifier(wxT("StereoToMono"));
+         const PluginDescriptor *plug = PluginManager::Get().GetPlugin(ID);
+         if (plug && plug->IsEnabled())
+            return Command( wxT("Stereo to Mono"),
+               XXO("Mix Stereo Down to &Mono"), OnStereoToMono,
+               AudioIONotBusyFlag() | StereoRequiredFlag() |
+                  WaveTracksSelectedFlag(), Options{} );
+         else
+            return {};
+      }
+   ) };
+   return menu;
+}
+
+AttachedItem sAttachment{
+   { wxT("Tracks/Mix/Mix"), { OrderingHint::Begin, {} } },
+   Shared( MenuItem() )
+};
+}
